Structure improvement.
Changed the main.py file and relatives libraries.

+++ project/libraries/flush_database.py
@@ -1,5 +1,10 @@
 import redis
 
 
-def clear_database(r: redis):
-    r.flushall()
+def flush_database(r: redis.Redis):
+    confirm = input("Are you sure you want to clear the database? This action cannot be undone. (y/n): ")
+    if confirm.lower() == 'y':
+        r.flushall()
+        print("Database cleared successfully!")
+    else:
+        print("Clear database operation canceled.")
